requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
fastapi==0.85.0
fonttools==4.37.4
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
torch==1.12.1
torchvision==0.13.1
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
